Separated the socket class into its own file, for maintainability
